app/api/deps.py
from fastapi import Depends
from sqlalchemy.orm import Session

from app.dependencies import get_db, get_current_user, get_current_active_user, get_current_active_superuser

# Re-export dependencies for simplicity
__all__ = ["get_db", "get_current_user", "get_current_active_user", "get_current_active_superuser"]
